CE-1400 DRY: remove compareResults method duplicate
